+++ src/frontend/app/shared-sdoc/components/sdoc-list-item-page/sdoc-list-item-page.component.ts
@@ -7,6 +7,10 @@ import {ActionTagEvent} from '../sdoc-actiontags/sdoc-actiontags.component';
 import {AngularHtmlService} from '../../../../shared/angular-commons/services/angular-html.service';
 import {AngularMarkdownService} from '../../../../shared/angular-commons/services/angular-markdown.service';
 import {PlatformService} from '../../../../shared/angular-commons/services/platform.service';
+import {SDocSearchResult} from '../../../../shared/sdoc-commons/model/container/sdoc-searchresult';
+import {SDocSearchForm} from '../../../../shared/sdoc-commons/model/forms/sdoc-searchform';
+import {SDocDataService} from '../../../../shared/sdoc-commons/services/sdoc-data.service';
+import {SDocSearchFormConverter} from '../../services/sdoc-searchform-converter.service';
 
 @Component({
     selector: 'app-sdoc-list-item-page',
@@ -53,7 +57,8 @@ export class SDocListItemPageComponent implements OnChanges {
     public showImage: EventEmitter<SDocRecord> = new EventEmitter();
 
     constructor(contentUtils: SDocContentUtils, private cd: ChangeDetectorRef, private platformService: PlatformService,
-                private angularMarkdownService: AngularMarkdownService, private angularHtmlService: AngularHtmlService) {
+                private angularMarkdownService: AngularMarkdownService, private angularHtmlService: AngularHtmlService,
+                private sdocDataService: SDocDataService, private searchFormConverter: SDocSearchFormConverter) {
         this.contentUtils = contentUtils;
     }
 
@@ -97,6 +102,8 @@ export class SDocListItemPageComponent implements OnChanges {
     }
 
     private updateData() {
+        const me = this;
+
         this.contentUtils.updateItemData(this.listItem, this.record, 'page');
         this.maxImageHeight = (window.innerHeight - 150) + 'px';
         if (this.record.type === 'IMAGE' || this.record.type === 'VIDEO') {
@@ -104,5 +111,25 @@ export class SDocListItemPageComponent implements OnChanges {
             this.listItem.flgShowProfileMap = this.listItem.flgShowProfileMap &&  this.imageShowMap;
         }
         this.cd.markForCheck();
+
+        if (this.record.type === 'TRIP' && (me.listItem.tracks === undefined || me.listItem.tracks.length < 1)) {
+            const searchForm = new SDocSearchForm({});
+            this.searchFormConverter.paramsToSearchForm(this.contentUtils.getSDocSubItemFiltersForType(this.record, 'TRACK', undefined),
+                {}, searchForm);
+            this.sdocDataService.search(searchForm, {
+                showFacets: false,
+                loadTrack: true,
+                showForm: false
+            }).then(function doneSearch(sdocSearchResult: SDocSearchResult) {
+                if (sdocSearchResult !== undefined && sdocSearchResult.recordCount > 0 &&
+                    sdocSearchResult.currentRecords[0].tripId === me.listItem.currentRecord.tripId) {
+                    // open map only if tripId of searchResut is for current trip: because of async
+                    me.listItem.flgShowMap = sdocSearchResult.recordCount > 0;
+                    me.listItem.flgShowProfileMap = me.listItem.flgShowMap;
+                    me.listItem.tracks = sdocSearchResult.currentRecords;
+                }
+                me.cd.markForCheck();
+            });
+        }
     }
 }